Updated the manual extension
